+++ cookbooks/iptables/recipes/default.rb
@@ -17,7 +17,13 @@
 # limitations under the License.
 #
 
-package "iptables" 
+
+
+if platform_family?("rhel") && node["platform_version"].to_i == 7
+  package "iptables-services"
+else
+  package "iptables"
+end
 
 execute "rebuild-iptables" do
   command "/usr/sbin/rebuild-iptables"
@@ -47,6 +53,9 @@ when "ubuntu", "debian"
   end
 end
 
-
-iptables_rule "all_established"
-iptables_rule "all_icmp"
+if node["iptables"]["install_rules"]
+  iptables_rule "all_established"
+  iptables_rule "all_icmp"
+  iptables_rule "prefix"
+  iptables_rule "postfix"
+end