+++ cookbooks/logrotate/definitions/logrotate_app.rb
@@ -35,14 +35,16 @@ log_rotate_params = {
 define(:logrotate_app, log_rotate_params) do
   include_recipe 'logrotate::default'
 
-  acceptable_options = %w(missingok compress delaycompress dateext dateyesterday copytruncate notifempty delaycompress ifempty mailfirst nocompress nocopy nocopytruncate nocreate nodelaycompress nomail nomissingok noolddir nosharedscripts notifempty sharedscripts)
+  acceptable_options = %w(missingok compress delaycompress dateext dateyesterday copy copytruncate notifempty delaycompress ifempty mailfirst maillast nocompress nocopy nocopytruncate nocreate nodelaycompress nomail nomissingok noolddir nosharedscripts noshred notifempty sharedscripts shred nodateext)
   options_tmp = params[:options] ||= %w(missingok compress delaycompress copytruncate notifempty)
   options = options_tmp.respond_to?(:each) ? options_tmp : options_tmp.split
 
   if params[:enable]
     invalid_options = options - acceptable_options
+
     unless invalid_options.empty?
-      Chef::Application.fatal! "The passed value(s) [#{invalid_options.join(',')}] are not valid"
+      Chef::Log.error("Invalid option(s) passed to logrotate: #{invalid_options.join(', ')}")
+      raise
     end
 
     template "/etc/logrotate.d/#{params[:name]}" do
@@ -53,22 +55,32 @@ define(:logrotate_app, log_rotate_params) do
       group    params[:template_group]
       backup   false
       variables(
-        :path          => Array(params[:path]).map { |path| path.to_s.inspect }.join(' '),
-        :create        => params[:create],
-        :frequency     => params[:frequency],
-        :dateformat    => params[:dateformat],
-        :size          => params[:size],
-        :minsize       => params[:minsize],
-        :maxsize       => params[:maxsize],
-        :su            => params[:su],
-        :rotate        => params[:rotate],
-        :olddir        => params[:olddir],
-        :sharedscripts => params[:sharedscripts],
-        :postrotate    => Array(params[:postrotate]).join("\n"),
-        :prerotate     => Array(params[:prerotate]).join("\n"),
-        :firstaction   => Array(params[:firstaction]).join("\n"),
-        :lastaction    => Array(params[:lastaction]).join("\n"),
-        :options       => options
+        :path             => Array(params[:path]).map { |path| path.to_s.inspect }.join(' '),
+        :create           => params[:create],
+        :frequency        => params[:frequency],
+        :dateformat       => params[:dateformat],
+        :extension        => params[:extension],
+        :tabooext         => params[:tabooext],
+        :mail             => params[:mail],
+        :size             => params[:size],
+        :minsize          => params[:minsize],
+        :maxsize          => params[:maxsize],
+        :maxage           => params[:maxage],
+        :su               => params[:su],
+        :rotate           => params[:rotate],
+        :start            => params[:start],
+        :olddir           => params[:olddir],
+        :compresscmd      => params[:compresscmd],
+        :compressoptions  => params[:compressoptions],
+        :uncompresscmd    => params[:uncompresscmd],
+        :compressext      => params[:compressext],
+        :shredcycles      => params[:shredcycles],
+        :sharedscripts    => params[:sharedscripts],
+        :postrotate       => Array(params[:postrotate]).join("\n"),
+        :prerotate        => Array(params[:prerotate]).join("\n"),
+        :firstaction      => Array(params[:firstaction]).join("\n"),
+        :lastaction       => Array(params[:lastaction]).join("\n"),
+        :options          => options
       )
     end
   else