+++ cookbooks/mysql/spec/unit/mysql_server/amazon/2014.03/custom_stepinto_spec.rb
@@ -41,14 +41,14 @@ GRANT ALL PRIVILEGES ON *.* TO 'root'@'1.2.3.4/5' IDENTIFIED BY 'YUNOSETPASSWORD
   context 'when using default parameters' do
     it 'creates mysql_service[amazon_2014_03_custom]' do
       expect(amazon_2014_03_custom_run).to create_mysql_service('amazon_2014_03_custom').with(
-        :version => '5.5',
-        :port => '3308',
-        :data_dir => '/data'
+        :parsed_version => '5.5',
+        :parsed_port => '3308',
+        :parsed_data_dir => '/data'
         )
     end
 
-    it 'steps into mysql_service and installs package[community-mysql-server]' do
-      expect(amazon_2014_03_custom_run).to install_package('mysql-server')
+    it 'steps into mysql_service and installs package[mysql-community-server]' do
+      expect(amazon_2014_03_custom_run).to install_package('mysql-community-server')
     end
 
     it 'steps into mysql_service and creates directory[/etc/mysql/conf.d]' do