+++ cookbooks/mysql/spec/unit/mysql_server/amazon/2014.03/default_stepinto_spec.rb
@@ -21,7 +21,7 @@ socket                         = /var/lib/mysql/mysql.sock
 
 [mysqld]
 user                           = mysql
-pid-file                       = /var/run/mysql/mysql.pid
+pid-file                       = /var/run/mysqld/mysql.pid
 socket                         = /var/lib/mysql/mysql.sock
 port                           = 3306
 datadir                        = /var/lib/mysql
@@ -48,14 +48,14 @@ SET PASSWORD FOR 'root'@'127.0.0.1' = PASSWORD('ilikerandompasswords');"
   context 'when using default parameters' do
     it 'creates mysql_service[amazon_2014_03_default]' do
       expect(amazon_2014_03_default_run).to create_mysql_service('amazon_2014_03_default').with(
-        :version => '5.5',
-        :port => '3306',
-        :data_dir => '/var/lib/mysql'
+        :parsed_version => '5.5',
+        :parsed_port => '3306',
+        :parsed_data_dir => '/var/lib/mysql'
         )
     end
 
     it 'steps into mysql_service and installs package[community-mysql-server]' do
-      expect(amazon_2014_03_default_run).to install_package('mysql-server')
+      expect(amazon_2014_03_default_run).to install_package('mysql-community-server')
     end
 
     it 'steps into mysql_service and creates directory[/etc/mysql/conf.d]' do