+++ cookbooks/mysql/spec/unit/mysql_server/ubuntu/10.04/default_stepinto_spec.rb
@@ -48,9 +48,9 @@ SET PASSWORD FOR 'root'@'127.0.0.1' = PASSWORD('ilikerandompasswords');"
   context 'when using default parameters' do
     it 'creates mysql_service[ubuntu_10_04_default]' do
       expect(ubuntu_10_04_default_run).to create_mysql_service('ubuntu_10_04_default').with(
-        :version => '5.1',
-        :port => '3306',
-        :data_dir => '/var/lib/mysql'
+        :parsed_version => '5.1',
+        :parsed_port => '3306',
+        :parsed_data_dir => '/var/lib/mysql'
         )
     end
 
@@ -81,8 +81,8 @@ SET PASSWORD FOR 'root'@'127.0.0.1' = PASSWORD('ilikerandompasswords');"
         )
     end
 
-    it 'steps into mysql_service and installs package[mysql-server]' do
-      expect(ubuntu_10_04_default_run).to install_package('mysql-server')
+    it 'steps into mysql_service and installs package[mysql-server-5.1]' do
+      expect(ubuntu_10_04_default_run).to install_package('mysql-server-5.1')
     end
 
     # apparmor